Separate file for DBPinnableSlice
