+++ src/db.rs
@@ -15,8 +15,8 @@
 
 use crate::{
     ffi, ffi_util::opt_bytes_to_ptr, ColumnFamily, ColumnFamilyDescriptor, DBIterator,
-    DBRawIterator, DBWALIterator, Direction, Error, FlushOptions, IteratorMode, Options,
-    WriteOptions, DB,
+    DBPinnableSlice, DBRawIterator, DBWALIterator, Direction, Error, FlushOptions, IteratorMode,
+    Options, WriteOptions, DB,
 };
 
 use libc::{self, c_char, c_int, c_uchar, c_void, size_t};
@@ -24,14 +24,18 @@ use std::collections::BTreeMap;
 use std::ffi::{CStr, CString};
 use std::fmt;
 use std::fs;
-use std::marker::PhantomData;
-use std::ops::Deref;
 use std::path::Path;
 use std::ptr;
 use std::slice;
 use std::str;
 
+// Safety note: auto-implementing Send on most db-related types is prevented by the inner FFI
+// pointer. In most cases, however, this pointer is Send-safe because it is never aliased and
+// rocksdb internally does not rely on thread-local information for its user-exposed types.
 unsafe impl Send for DB {}
+
+// Sync is similarly safe for many types because they do not expose interior mutability, and their
+// use within the rocksdb library is generally behind a const reference
 unsafe impl Sync for DB {}
 
 #[derive(Debug, Copy, Clone, PartialEq)]
@@ -1423,13 +1427,8 @@ impl Default for ReadOptions {
     }
 }
 
-// Safety note: auto-implementing Send on most db-related types is prevented by the inner FFI
-// pointer. In most cases, however, this pointer is Send-safe because it is never aliased and
-// rocksdb internally does not rely on thread-local information for its user-exposed types.
 unsafe impl Send for ReadOptions {}
 
-// Sync is similarly safe for many types because they do not expose interior mutability, and their
-// use within the rocksdb library is generally behind a const reference
 unsafe impl Sync for ReadOptions {}
 
 fn to_cpath<P: AsRef<Path>>(path: P) -> Result<CString, Error> {
@@ -1439,65 +1438,6 @@ fn to_cpath<P: AsRef<Path>>(path: P) -> Result<CString, Error> {
             "Failed to convert path to CString: {}",
             e,
         ))),
-    }
-}
-
-/// Wrapper around RocksDB PinnableSlice struct.
-///
-/// With a pinnable slice, we can directly leverage in-memory data within
-/// RocksDB to avoid unnecessary memory copies. The struct here wraps the
-/// returned raw pointer and ensures proper finalization work.
-pub struct DBPinnableSlice<'a> {
-    ptr: *mut ffi::rocksdb_pinnableslice_t,
-    db: PhantomData<&'a DB>,
-}
-
-// Safety note: auto-implementing Send on most db-related types is prevented by the inner FFI
-// pointer. In most cases, however, this pointer is Send-safe because it is never aliased and
-// rocksdb internally does not rely on thread-local information for its user-exposed types.
-unsafe impl<'a> Send for DBPinnableSlice<'a> {}
-
-// Sync is similarly safe for many types because they do not expose interior mutability, and their
-// use within the rocksdb library is generally behind a const reference
-unsafe impl<'a> Sync for DBPinnableSlice<'a> {}
-
-impl<'a> AsRef<[u8]> for DBPinnableSlice<'a> {
-    fn as_ref(&self) -> &[u8] {
-        // Implement this via Deref so as not to repeat ourselves
-        &*self
-    }
-}
-
-impl<'a> Deref for DBPinnableSlice<'a> {
-    type Target = [u8];
-
-    fn deref(&self) -> &[u8] {
-        unsafe {
-            let mut val_len: size_t = 0;
-            let val = ffi::rocksdb_pinnableslice_value(self.ptr, &mut val_len) as *mut u8;
-            slice::from_raw_parts(val, val_len)
-        }
-    }
-}
-
-impl<'a> Drop for DBPinnableSlice<'a> {
-    fn drop(&mut self) {
-        unsafe {
-            ffi::rocksdb_pinnableslice_destroy(self.ptr);
-        }
-    }
-}
-
-impl<'a> DBPinnableSlice<'a> {
-    /// Used to wrap a PinnableSlice from rocksdb to avoid unnecessary memcpy
-    ///
-    /// # Unsafe
-    /// Requires that the pointer must be generated by rocksdb_get_pinned
-    unsafe fn from_c(ptr: *mut ffi::rocksdb_pinnableslice_t) -> DBPinnableSlice<'a> {
-        DBPinnableSlice {
-            ptr,
-            db: PhantomData,
-        }
     }
 }
